policycoreutils: Add test for SCTP support in semanage
This test simulates a use case where a user wants to confine a
client-server application communicating over SCTP protocol.

A custom policy module is used to introduce a new port type, domain type
for the application and an executable file type for the application
executable as well as a minimal set of allow rules for sctp_test to work
properly.
"semanage port" is then used to define SCTP contexts for ports 1025 and
1026.
Lastly sctp_test is executed in client-server configuration.