policycoreutils: Add test for restorecond "blob data"
This test creates mislabeled files and runs restorecond repeatedly to test
for an issue caused by overwriting program variable with memory address.
The issue caused restorecond to write status messages (that appeared
as "blob data" because of \r character) to journal.

Since the memory address mentioned above is effectively random data,
this test may result in SUCESS even if the issue being tested is present!